+++ src/main/java/blusunrize/immersiveengineering/client/IEManual.java
@@ -33,6 +33,8 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.JSONUtils;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.world.dimension.DimensionType;
 import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.fml.ModLoadingContext;
@@ -265,7 +267,7 @@ public class IEManual
 	)
 	{
 		return () -> {
-			String[][] table = formatToTable_ItemIntMap(getContents.get(), valueType);
+			ITextComponent[][] table = formatToTable_ItemIntMap(getContents.get(), valueType);
 			return new ManualElementTable(ManualHelper.getManual(), table, false);
 		};
 	}
@@ -381,16 +383,16 @@ public class IEManual
 		return builder.create();
 	}
 
-	static <T> String[][] formatToTable_ItemIntMap(Map<T, Integer> map, String valueType)
+	static <T> ITextComponent[][] formatToTable_ItemIntMap(Map<T, Integer> map, String valueType)
 	{
 		List<Entry<T, Integer>> sortedMapArray = new ArrayList<>(map.entrySet());
 		sortedMapArray.sort(Comparator.comparing(Entry::getValue));
-		ArrayList<String[]> list = new ArrayList<>();
+		ArrayList<ITextComponent[]> list = new ArrayList<>();
 		try
 		{
 			for(Entry<T, Integer> entry : sortedMapArray)
 			{
-				String item = entry.getKey().toString();
+				ITextComponent item = null;
 				if(entry.getKey() instanceof ResourceLocation)
 				{
 					ResourceLocation key = (ResourceLocation)entry.getKey();
@@ -398,20 +400,21 @@ public class IEManual
 					{
 						ItemStack is = IEApi.getPreferredTagStack(key);
 						if(!is.isEmpty())
-							item = is.getDisplayName().getFormattedText();
+							item = is.getDisplayName();
 					}
 				}
+				else if(entry.getKey() instanceof ITextComponent)
+					item = (ITextComponent)entry.getKey();
+				if(item==null)
+					item = new StringTextComponent(entry.getKey().toString());
 
-				if(item!=null)
-				{
-					int bt = entry.getValue();
-					String am = bt+" "+valueType;
-					list.add(new String[]{item, am});
-				}
+				int bt = entry.getValue();
+				ITextComponent am = new StringTextComponent(bt+" "+valueType);
+				list.add(new ITextComponent[]{item, am});
 			}
 		} catch(Exception e)
 		{
 		}
-		return list.toArray(new String[0][]);
+		return list.toArray(new ITextComponent[0][]);
 	}
 }