+++ src/main/java/blusunrize/immersiveengineering/common/blocks/IEBlockInterfaces.java
@@ -302,6 +302,11 @@ public class IEBlockInterfaces
 		boolean hammerUseSide(Direction side, PlayerEntity player, Vec3d hitVec);
 	}
 
+	public interface IScrewdriverInteraction
+	{
+		boolean screwdriverUseSide(Direction side, PlayerEntity player, Vec3d hitVec);
+	}
+
 	public interface IPlacementInteraction
 	{
 		void onTilePlaced(World world, BlockPos pos, BlockState state, Direction side, float hitX, float hitY, float hitZ, LivingEntity placer, ItemStack stack);
@@ -452,11 +457,6 @@ public class IEBlockInterfaces
 		int[] getCurrentProcessesStep();
 
 		int[] getCurrentProcessesMax();
-	}
-
-	public interface INeighbourChangeTile
-	{
-		void onNeighborBlockChange(BlockPos otherPos);
 	}
 
 	public interface IPropertyPassthrough