+++ src/main/java/blusunrize/immersiveengineering/common/blocks/generic/MultiblockPartTileEntity.java
@@ -15,6 +15,7 @@ import blusunrize.immersiveengineering.common.blocks.IEBaseTileEntity;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.*;
 import blusunrize.immersiveengineering.common.blocks.multiblocks.IETemplateMultiblock;
 import blusunrize.immersiveengineering.common.util.ChatUtils;
+import blusunrize.immersiveengineering.common.util.SafeChunkUtils;
 import blusunrize.immersiveengineering.common.util.Utils;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
@@ -46,7 +47,7 @@ import java.util.Optional;
 import java.util.Set;
 
 public abstract class MultiblockPartTileEntity<T extends MultiblockPartTileEntity<T>> extends IEBaseTileEntity
-		implements ITickableTileEntity, IDirectionalTile, IGeneralMultiblock, IHammerInteraction, IMirrorAble
+		implements ITickableTileEntity, IDirectionalTile, IGeneralMultiblock, IScrewdriverInteraction, IMirrorAble
 {
 	public boolean formed = false;
 	//Position of this block according to the BlockInfo's returned by IMultiblock#getStructure
@@ -374,7 +375,7 @@ public abstract class MultiblockPartTileEntity<T extends MultiblockPartTileEntit
 	}
 
 	@Override
-	public boolean hammerUseSide(Direction side, PlayerEntity player, Vec3d hitVec)
+	public boolean screwdriverUseSide(Direction side, PlayerEntity player, Vec3d hitVec)
 	{
 		if(!world.isRemote)
 		{
@@ -406,7 +407,7 @@ public abstract class MultiblockPartTileEntity<T extends MultiblockPartTileEntit
 			T tile = this.getTileForPos(rsPos);
 			if(tile!=null)
 			{
-				boolean b = getWorldNonnull().getRedstonePowerFromNeighbors(tile.getPos()) > 0;
+				boolean b = tile.isRSPowered();
 				if(redstoneControlInverted!=b)
 					return true;
 			}
@@ -418,7 +419,7 @@ public abstract class MultiblockPartTileEntity<T extends MultiblockPartTileEntit
 	public T getTileForPos(BlockPos targetPosInMB)
 	{
 		BlockPos target = getBlockPosForPos(targetPosInMB);
-		TileEntity tile = Utils.getExistingTileEntity(getWorldNonnull(), target);
+		TileEntity tile = SafeChunkUtils.getSafeTE(getWorldNonnull(), target);
 		if(this.getClass().isInstance(tile))
 			return (T)tile;
 		return null;