+++ src/main/java/blusunrize/lib/manual/ManualElementTable.java
@@ -9,25 +9,42 @@
 package blusunrize.lib.manual;
 
 import blusunrize.lib.manual.gui.ManualScreen;
+import com.google.common.base.Preconditions;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.gui.AbstractGui;
 import net.minecraft.client.gui.widget.button.Button;
-import net.minecraft.client.resources.I18n;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.OptionalInt;
 
 public class ManualElementTable extends SpecialManualElements
 {
-	private String[][] table;
-	private String[][] localizedTable;
+	private ITextComponent[][] table;
 	private int[] bars;
-	private boolean horizontalBars = false;
-	private int height;
+	private boolean horizontalBars;
+	private OptionalInt height = OptionalInt.empty();
 	private int[] textOff;
 
+	@Deprecated
 	public ManualElementTable(ManualInstance manual, String[][] table, boolean horizontalBars)
 	{
+		this(manual, Arrays.stream(table)
+						.map(a -> Arrays.stream(a)
+								.map(StringTextComponent::new)
+								.toArray(ITextComponent[]::new)
+						)
+						.toArray(ITextComponent[][]::new),
+				horizontalBars
+		);
+	}
+
+	public ManualElementTable(ManualInstance manual, ITextComponent[][] table, boolean horizontalBars)
+	{
 		super(manual);
+		Preconditions.checkNotNull(table);
 		this.table = table;
 		this.horizontalBars = horizontalBars;
 	}
@@ -38,55 +55,7 @@ public class ManualElementTable extends SpecialManualElements
 		super.onOpened(gui, x, y, pageButtons);
 		try
 		{
-			if(table!=null)
-			{
-				localizedTable = new String[table.length][];
-
-				bars = new int[1];
-				for(int i = 0; i < table.length; i++)
-				{
-					localizedTable[i] = new String[table[i].length];
-					for(int j = 0; j < table[i].length; j++)
-						if(table[i][j]!=null)
-							localizedTable[i][j] = I18n.format(table[i][j]);
-
-					if(table[i].length-1 > bars.length)
-					{
-						int[] newBars = new int[table[i].length-1];
-						System.arraycopy(bars, 0, newBars, 0, bars.length);
-						bars = newBars;
-					}
-					for(int j = 0; j < table[i].length-1; j++)
-					{
-						int fl = manual.fontRenderer().getStringWidth(localizedTable[i][j]);
-						if(fl > bars[j])
-							bars[j] = fl;
-					}
-				}
-				textOff = new int[bars!=null?bars.length: 0];
-				if(bars!=null)
-				{
-					int xx = x;
-					for(int i = 0; i < bars.length; i++)
-					{
-						xx += bars[i]+8;
-						textOff[i] = xx;
-					}
-				}
-
-				int yOff = 0;
-				for(int i = 0; i < localizedTable.length; i++)
-					if(localizedTable[i]!=null)
-						for(int j = 0; j < localizedTable[i].length; j++)
-							if(localizedTable[i][j]!=null)
-							{
-								int w = Math.max(10, 120-(j > 0?textOff[j-1]-x: 0));
-								int l = manual.fontRenderer().listFormattedStringToWidth(localizedTable[i][j], w).size();
-								if(j!=0)
-									yOff += l*(manual.fontRenderer().FONT_HEIGHT+1);
-							}
-				height = yOff;
-			}
+			recalculateLayout();
 		} catch(Exception e)
 		{
 			e.printStackTrace();
@@ -96,13 +65,13 @@ public class ManualElementTable extends SpecialManualElements
 	@Override
 	public void render(ManualScreen gui, int x, int y, int mx, int my)
 	{
-		if(localizedTable!=null)
+		if(table!=null)
 		{
 			int col = manual.getHighlightColour()|0xff000000;
 			AbstractGui.fill(x, y-2, x+120, y-1, col);
 
 			int yOff = 0;
-			for(String[] line : localizedTable)
+			for(ITextComponent[] line : table)
 				if(line!=null)
 				{
 					int height = 0;
@@ -111,8 +80,9 @@ public class ManualElementTable extends SpecialManualElements
 						{
 							int xx = textOff.length > 0&&j > 0?textOff[j-1]: x;
 							int w = Math.max(10, 120-(j > 0?textOff[j-1]-x: 0));
-							manual.fontRenderer().drawSplitString(line[j], xx, y+yOff, w, manual.getTextColour());
-							int lines = manual.fontRenderer().listFormattedStringToWidth(line[j], w).size();
+							String lineText = line[j].getFormattedText();
+							manual.fontRenderer().drawSplitString(lineText, xx, y+yOff, w, manual.getTextColour());
+							int lines = manual.fontRenderer().listFormattedStringToWidth(lineText, w).size();
 							if(lines > height)
 								height = lines;
 						}
@@ -142,10 +112,47 @@ public class ManualElementTable extends SpecialManualElements
 		return false;
 	}
 
+	private void recalculateLayout()
+	{
+		bars = new int[1];
+		for(ITextComponent[] tableLine : table)
+		{
+			if(tableLine.length-1 > bars.length)
+				bars = Arrays.copyOf(bars, tableLine.length-1);
+			for(int j = 0; j < tableLine.length-1; j++)
+			{
+				int fl = manual.fontRenderer().getStringWidth(tableLine[j].getFormattedText());
+				if(fl > bars[j])
+					bars[j] = fl;
+			}
+		}
+		textOff = new int[bars.length];
+		int xx = 0;
+		for(int i = 0; i < bars.length; i++)
+		{
+			xx += bars[i]+8;
+			textOff[i] = xx;
+		}
+		int yOff = 0;
+		for(ITextComponent[] tableLine : table)
+			if(tableLine!=null)
+				for(int j = 0; j < tableLine.length; j++)
+					if(tableLine[j]!=null)
+					{
+						int w = Math.max(10, 120-(j > 0?textOff[j-1]: 0));
+						int l = manual.fontRenderer().listFormattedStringToWidth(tableLine[j].getFormattedText(), w).size();
+						if(j!=0)
+							yOff += l*(manual.fontRenderer().FONT_HEIGHT+1);
+					}
+		height = OptionalInt.of(yOff);
+	}
+
 	@Override
 	public int getPixelsTaken()
 	{
-		return height;
+		if(!height.isPresent())
+			recalculateLayout();
+		return height.getAsInt();
 	}
 
 	@Override